build時のエラーを解決
